-Changed euclidian space to use doubles - Preliminary OBJ loading preparations - Added test OBJ file

+++ src/vector.h
@@ -13,7 +13,7 @@
 
 //Vector
 typedef struct {
-	float x, y, z;
+	double x, y, z;
 }vector;
 
 //Simulated light ray
